Adding More Responses

+++ IEXSharp/Model/PremiumData/WallStreetHorizon/Response/FiscalQuarterEndResponse.cs
@@ -7,7 +7,7 @@ namespace IEXSharp.Model.PremiumData.WallStreetHorizon.Response
 		public string companyname { get; set; }
 		public string quarterenddate { get; set; }
 		public string quarter { get; set; }
-		public int fiscalyear { get; set; }
+		public long fiscalyear { get; set; }
 		public string updated { get; set; }
 		public string id { get; set; }
 		public string source { get; set; }
